Move hasValidTimestamps
